requirements.txt
python-telegram-bot
python-telegram-bot-calendar
gspread
passlib
redis
google-api-python-client
google-auth-httplib2
google-auth-oauthlib